Moved line breaking functions to it's own file
